+++ test/airbrussh_test.rb
@@ -5,4 +5,15 @@ class AirbrusshTest < Minitest::Test
     config = Airbrussh.configuration
     assert_equal(config, Airbrussh.configure { |c| c })
   end
+
+  def test_configuration_returns_passed_config
+    config = Airbrussh::Configuration.new
+    assert_equal(config, Airbrussh.configuration(config))
+  end
+
+  def test_configuration_applies_options
+    config = Airbrussh.configuration(:banner => "test_success")
+    assert_equal("test_success", config.banner)
+    assert_equal("test_success", Airbrussh.configuration.banner)
+  end
 end